Module wasmer_runtime_core::backend

Modules

sys

Structs

BackendCompilerConfig
Use this to point to a compiler config struct provided by the backend. The backend struct must support runtime reflection with Any, which is any struct that does not contain a non-'static reference.

CompilerConfig
Configuration data for the compiler

ExceptionTable
An exception table for a RunnableModule.

Features
Controls which experimental features will be enabled. Features usually have a corresponding WebAssembly proposal.

InlineBreakpoint
Information of an inline breakpoint.

SigRegistry
The SigRegistry represents a process-global map of function signatures to signature indexes and vice versa (the map goes both ways).

Token
This type cannot be constructed from outside the runtime crate.

Enums

Architecture
The target architecture for code generation.

ExceptionCode
The code of an exception.

InlineBreakpointType
The type of an inline breakpoint.

MemoryBoundCheckMode

Constants

_IMPL_DESERIALIZE_FOR_ExceptionCode

_IMPL_DESERIALIZE_FOR_ExceptionTable

_IMPL_SERIALIZE_FOR_ExceptionCode

_IMPL_SERIALIZE_FOR_ExceptionTable

Traits

CacheGen

Compiler

RunnableModule