Trait wasmer_runtime_core::backend::RunnableModule

pub trait RunnableModule: Send + Sync {
    fn get_func(
        &self, 
        info: &ModuleInfo, 
        local_func_index: LocalFuncIndex
    ) -> Option<NonNull<Func>>;

    fn get_trampoline(
        &self, 
        info: &ModuleInfo, 
        sig_index: SigIndex
    ) -> Option<Wasm>;

    unsafe fn do_early_trap(&self, data: RuntimeError) -> !;

    fn get_module_state_map(&self) -> Option<ModuleStateMap> { ... }

    fn get_breakpoints(&self) -> Option<BreakpointMap> { ... }

    fn get_exception_table(&self) -> Option<&ExceptionTable> { ... }

    unsafe fn patch_local_function(
        &self, 
        _idx: usize, 
        _target_address: usize
    ) -> bool { ... }

    fn get_code(&self) -> Option<&[u8]> { ... }

    fn get_offsets(&self) -> Option<Vec<usize>> { ... }

    fn get_local_function_offsets(&self) -> Option<Vec<usize>> { ... }

    fn get_inline_breakpoint_size(&self, _arch: Architecture) -> Option<usize> { ... }

    fn read_inline_breakpoint(
        &self, 
        _arch: Architecture, 
        _code: &[u8]
    ) -> Option<InlineBreakpoint> { ... }
}

Required methods

fn get_func(
&self,
info: &ModuleInfo,
local_func_index: LocalFuncIndex
) -> Option<NonNull<Func>>

This returns a pointer to the function designated by the local_func_index parameter.

fn get_trampoline(&self, info: &ModuleInfo, sig_index: SigIndex) -> Option<Wasm>

A wasm trampoline contains the necessary data to dynamically call an exported wasm function. Given a particular signature index, we return a trampoline that is matched with that signature and an invoke function that can call the trampoline.

unsafe fn do_early_trap(&self, data: RuntimeError) -> !

Trap an error.

Loading content...
Provided methods

fn get_module_state_map(&self) -> Option<ModuleStateMap>

fn get_breakpoints(&self) -> Option<BreakpointMap>

fn get_exception_table(&self) -> Option<&ExceptionTable>

unsafe fn patch_local_function(
&self,
_idx: usize,
_target_address: usize
) -> bool

fn get_code(&self) -> Option<&[u8]>

Returns the machine code associated with this module.

fn get_offsets(&self) -> Option<Vec<usize>>

Returns the beginning offsets of all functions, including import trampolines.

fn get_local_function_offsets(&self) -> Option<Vec<usize>>

Returns the beginning offsets of all local functions.

fn get_inline_breakpoint_size(&self, _arch: Architecture) -> Option<usize>

Returns the inline breakpoint size corresponding to an Architecture (None in case is not implemented)

fn read_inline_breakpoint(
&self,
_arch: Architecture,
_code: &[u8]
) -> Option<InlineBreakpoint>

Attempts to read an inline breakpoint from the code.

Inline breakpoints are detected by special instruction sequences that never appear in valid code.

Loading content...