Module wasmer_runtime_core::cache

The cache module provides the common data structures used by compiler backends to allow serializing compiled wasm code to a binary format. The binary format can be persisted, and loaded to allow skipping compilation and fast startup.

Structs

Artifact
Artifact are produced by caching, are serialized/deserialized to binaries, and contain module info, backend metadata, and compiled code.

WasmHash
The hash of a wasm module.

Enums

Error
Kinds of caching errors

InvalidFileType
Indicates the invalid type of invalid cache file

Constants

WASMER_VERSION_HASH
A unique ID generated from the version of Wasmer for use with cache versioning