Struct wasmer_runtime_core::cache::WasmHash

pub struct WasmHash([u8; 32]);

The hash of a wasm module.

Used as a key when loading and storing modules in a Cache.

Methods

impl WasmHash

pub fn generate(wasm: &[u8]) -> Self

Hash a wasm module.

Note:

This does no verification that the supplied data is, in fact, a wasm module.

pub fn encode(self) -> String

Create the hexadecimal representation of the stored hash.

pub fn decode(hex_str: &str) -> Result<Self, Error>

Create hash from hexadecimal representation

pub(crate) fn into_array(self) -> [u8; 32]