Module wasmer_runtime_core::export

The export module contains the implementation data structures and helper functions used to manipulate and access a wasm module's exports including memories, tables, globals, and functions.

Structs

ExportIter
An iterator to an instance's exports.

FuncPointer
Const pointer to a Func.

Enums

Context
A kind of Context.

Export
Kind of WebAssembly export.