Module wasmer_runtime_core::global

The global module contains the implementation data structures and helper functions used to manipulate and access a wasm globals.

Structs

Global
Container with a descriptor and a reference to a global value.