Module wasmer_runtime_core::import

The import module contains the implementation data structures and helper functions used to manipulate and access a wasm module's imports including memories, tables, globals, and functions.

Structs

ImportObject
All of the import data used when instantiating.

ImportObjectIterator
Iterator for an ImportObject's exports.

Namespace
The top-level container for the two-level wasm imports

Traits

IsExport
A trait that represents Export values.

LikeNamespace
This trait represents objects that act as a namespace for imports. For example, an Instance or ImportObject could be considered namespaces that could provide imports to an instance.