Struct wasmer_runtime_core::instance::Exports

pub struct Exports {
    instance_inner: *const InstanceInner,
    module: Arc<ModuleInner>,
}

Exports is used to get exports like Funcs, DynFuncs, Memorys, Globals, and Tables from an Instance.

Use Instance.exports to get an Exports from an Instance.

Fields
instance_inner: *const InstanceInnermodule: Arc<ModuleInner>
Methods

impl Exports

pub fn get<'a, T: Exportable<'a>>(&'a self, name: &str) -> ResolveResult<T>

Get an export.

// We can get a function as a static `Func`
let func: Func<i32, i32> = instance.exports.get("my_func")?;
let _result = func.call(42);

// Or we can get it as a dynamic `DynFunc`
let dyn_func: DynFunc = instance.exports.get("my_func")?;
let _result= dyn_func.call(&[Value::I32(42)]);

// We can also get other exports like `Global`s, `Memory`s, and `Table`s
let _counter: Global = instance.exports.get("counter")?;

fn get_inner(&self) -> (&InstanceInner, &ModuleInner)

This method must remain private for Exports to be sound.

Important traits for ExportIter<'a>
impl<'a> Iterator for ExportIter<'a> type Item = (String, Export);

pub fn into_iter(&self) -> ExportIter

Iterate the exports.

for (export_name, export_value) in instance.exports.into_iter() {
   println!("Found export `{}` with value `{:?}`", export_name, export_value);
}