Module wasmer_runtime_core::jit_debug

Code for interacting with the GDB JIT interface.

Structs

JIT_DEBUG_DESCRIPTOR_LOCK
Global lock on __jit_debug_descriptor. Acquire this lock when reading or writing to the global variable. This includes calls to __jit_debug_register_code which may cause a debugger to read from the global variable.

JitCodeDebugInfoEntryHandle
Handle to debug info about JIT code registered with a debugger

JitCodeDebugInfoEntryHandleInner
Type for implementing Drop on the memory shared with the debugger.

JitCodeDebugInfoManager
Manager of debug info registered with the debugger.

JitCodeEntry
Node of the doubly linked list that the GDB JIT interface reads from.

JitDebugDescriptor
Head node of the doubly linked list that the GDB JIT interface expects.

Enums

JitAction
The operation that the debugger should perform with the entry that we gave it.

Statics

JIT_DEBUG_DESCRIPTOR_LOCK

__jit_debug_descriptor

Functions

__jit_debug_register_code⚠

push_front⚠
Prepend an item to the front of the __jit_debug_descriptor entry list

remove_node⚠
Removes an entry from the doubly linked list, updating both nodes that it's connected to.