Struct wasmer_runtime_core::jit_debug::JIT_DEBUG_DESCRIPTOR_LOCK

struct JIT_DEBUG_DESCRIPTOR_LOCK {
    __private_field: (),
}

Global lock on __jit_debug_descriptor. Acquire this lock when reading or writing to the global variable. This includes calls to __jit_debug_register_code which may cause a debugger to read from the global variable.

Fields
__private_field: ()