Trait wasmer_runtime_core::loader::Instance

pub trait Instance {
    type Error: Debug;
    fn call(&mut self, id: usize, args: &[Value]) -> Result<u128, Self::Error>;

    fn read_memory(
        &mut self, 
        _offset: u32, 
        _len: u32
    ) -> Result<Vec<u8>, Self::Error> { ... }

    fn write_memory(
        &mut self, 
        _offset: u32, 
        _len: u32, 
        _buf: &[u8]
    ) -> Result<(), Self::Error> { ... }
}

This trait represents an instance used by the loader.

Associated Types

type Error: Debug

The error type returned by this instance.

Loading content...
Required methods

fn call(&mut self, id: usize, args: &[Value]) -> Result<u128, Self::Error>

Call a function by id with the given args.

Loading content...
Provided methods

fn read_memory(
&mut self,
_offset: u32,
_len: u32
) -> Result<Vec<u8>, Self::Error>

Read memory at the given offset and length.

fn write_memory(
&mut self,
_offset: u32,
_len: u32,
_buf: &[u8]
) -> Result<(), Self::Error>

Write memory at the given offset and length.

Loading content...