Trait wasmer_runtime_core::loader::Loader

pub trait Loader {
    type Instance: Instance;
    type Error: Debug;
    fn load(
        &self, 
        rm: &dyn RunnableModule, 
        module: &ModuleInfo, 
        ctx: &Ctx
    ) -> Result<Self::Instance, Self::Error>;
}

The loader trait represents the functions used to load an instance.

Associated Types

type Instance: Instance

The type of Instance for the loader.

type Error: Debug

The error type returned by the loader.

Loading content...
Required methods

fn load(
&self,
rm: &dyn RunnableModule,
module: &ModuleInfo,
ctx: &Ctx
) -> Result<Self::Instance, Self::Error>

Loads the given module and context into an instance.

Loading content...