Module wasmer_runtime_core::memory

The memory module contains the implementation data structures and helper functions used to manipulate and access wasm memory.

Modules

ptr
A reusable pointer abstraction for getting memory from the guest's memory.

Structs

Atomically
Atomically.

DynamicMemory
This is an internal-only api.

Memory
A shared or unshared wasm linear memory.

MemoryView
A view into a memory.

SharedMemory
A reference to a shared memory.

SharedMemoryInternal
Data structure for a shared internal memory.

StaticMemory
This is an internal-only api.

UnsharedMemory
A reference to an unshared memory.

Enums

MemoryType
A kind a memory.