Struct wasmer_runtime_core::memory::ptr::WasmPtr

#[repr(transparent)]
pub struct WasmPtr<T: Copy, Ty = Item> {
    offset: u32,
    _phantom: PhantomData<(T, Ty)>,
}

A zero-cost type that represents a pointer to something in Wasm linear memory.

This type can be used directly in the host function arguments:

pub fn host_import(ctx: &mut Ctx, ptr: WasmPtr<u32>) {
    let memory = ctx.memory(0);
    let derefed_ptr = ptr.deref(memory).expect("pointer in bounds");
    let inner_val: u32 = derefed_ptr.get();
    println!("Got {} from Wasm memory address 0x{:X}", inner_val, ptr.offset());
    // update the value being pointed to
    derefed_ptr.set(inner_val + 1);
}

Fields
offset: u32_phantom: PhantomData<(T, Ty)>
Methods

impl<T: Copy, Ty> WasmPtr<T, Ty>

Methods relevant to all types of WasmPtr.

pub fn new(offset: u32) -> Self

Create a new WasmPtr at the given offset.

pub fn offset(self) -> u32

Get the offset into Wasm linear memory for this WasmPtr.

impl<T: Copy + ValueType> WasmPtr<T, Item>

Methods for WasmPtrs to data that can be dereferenced, namely to types that implement ValueType, meaning that they're valid for all possible bit patterns.

pub fn deref<'a>(self, memory: &'a Memory) -> Option<&'a Cell<T>>

Dereference the WasmPtr getting access to a &Cell<T> allowing for reading and mutating of the inner value.

This method is unsound if used with unsynchronized shared memory. If you're unsure what that means, it likely does not apply to you. This invariant will be enforced in the future.

pub unsafe fn deref_mut<'a>(self, memory: &'a Memory) -> Option<&'a mut Cell<T>>

Mutably dereference this WasmPtr getting a &mut Cell<T> allowing for direct access to a &mut T.

Safety

This method does not do any aliasing checks: it's possible to create &mut T that point to the same memory. You should ensure that you have exclusive access to Wasm linear memory before calling this method.

impl<T: Copy + ValueType> WasmPtr<T, Array>

Methods for WasmPtrs to arrays of data that can be dereferenced, namely to types that implement ValueType, meaning that they're valid for all possible bit patterns.

pub fn deref(
self,
memory: &Memory,
index: u32,
length: u32
) -> Option<&[Cell<T>]>

Dereference the WasmPtr getting access to a &[Cell<T>] allowing for reading and mutating of the inner values.

This method is unsound if used with unsynchronized shared memory. If you're unsure what that means, it likely does not apply to you. This invariant will be enforced in the future.

pub unsafe fn deref_mut(
self,
memory: &Memory,
index: u32,
length: u32
) -> Option<&mut [Cell<T>]>

Mutably dereference this WasmPtr getting a &mut [Cell<T>] allowing for direct access to a &mut [T].

Safety

This method does not do any aliasing checks: it's possible to create &mut T that point to the same memory. You should ensure that you have exclusive access to Wasm linear memory before calling this method.

pub fn get_utf8_string(self, memory: &Memory, str_len: u32) -> Option<&str>

Get a UTF-8 string from the WasmPtr with the given length.

Note that this method returns a reference to Wasm linear memory. The underlying data can be mutated if the Wasm is allowed to execute or an aliasing WasmPtr is used to mutate memory.

pub fn get_utf8_string_with_nul(self, memory: &Memory) -> Option<&str>

Get a UTF-8 string from the WasmPtr, where the string is nul-terminated.

Note that this does not account for UTF-8 strings that contain nul themselves, [get_utf8_string] has to be used for those.

Also note that this method returns a reference to Wasm linear memory. The underlying data can be mutated if the Wasm is allowed to execute or an aliasing WasmPtr is used to mutate memory.