Module wasmer_runtime_core::module

The module module contains the implementation data structures and helper functions used to manipulate and access wasm modules.

Structs

DataInitializer
A data initializer for linear memory.

ImportName

Module
A compiled WebAssembly module.

ModuleInfo
Container for module data including memories, globals, tables, imports, and exports.

ModuleInner
This is used to instantiate a new WebAssembly module.

NameIndex
A type-safe handle referring to a name in a module namespace.

NamespaceIndex
A type-safe handle referring to a module namespace.

StringTable
A map of index to string.

StringTableBuilder
String table builder.

TableInitializer
A WebAssembly table initializer.

Enums

ExportIndex
A wrapper around the TypedIndexes for Wasm functions, Wasm memories, Wasm globals, and Wasm tables.

Constants

_IMPL_DESERIALIZE_FOR_DataInitializer

_IMPL_DESERIALIZE_FOR_ExportIndex

_IMPL_DESERIALIZE_FOR_ImportName

_IMPL_DESERIALIZE_FOR_ModuleInfo

_IMPL_DESERIALIZE_FOR_NameIndex

_IMPL_DESERIALIZE_FOR_NamespaceIndex

_IMPL_DESERIALIZE_FOR_StringTable

_IMPL_DESERIALIZE_FOR_TableInitializer

_IMPL_SERIALIZE_FOR_DataInitializer

_IMPL_SERIALIZE_FOR_ExportIndex

_IMPL_SERIALIZE_FOR_ImportName

_IMPL_SERIALIZE_FOR_ModuleInfo

_IMPL_SERIALIZE_FOR_NameIndex

_IMPL_SERIALIZE_FOR_NamespaceIndex

_IMPL_SERIALIZE_FOR_StringTable

_IMPL_SERIALIZE_FOR_TableInitializer