Struct wasmer_runtime_core::module::ModuleInfo

pub struct ModuleInfo {
    pub memories: Map<LocalMemoryIndex, MemoryDescriptor>,
    pub globals: Map<LocalGlobalIndex, GlobalInit>,
    pub tables: Map<LocalTableIndex, TableDescriptor>,
    pub imported_functions: Map<ImportedFuncIndex, ImportName>,
    pub imported_memories: Map<ImportedMemoryIndex, (ImportName, MemoryDescriptor)>,
    pub imported_tables: Map<ImportedTableIndex, (ImportName, TableDescriptor)>,
    pub imported_globals: Map<ImportedGlobalIndex, (ImportName, GlobalDescriptor)>,
    pub exports: IndexMap<String, ExportIndex>,
    pub data_initializers: Vec<DataInitializer>,
    pub elem_initializers: Vec<TableInitializer>,
    pub start_func: Option<FuncIndex>,
    pub func_assoc: Map<FuncIndex, SigIndex>,
    pub signatures: Map<SigIndex, FuncSig>,
    pub backend: String,
    pub namespace_table: StringTable<NamespaceIndex>,
    pub name_table: StringTable<NameIndex>,
    pub em_symbol_map: Option<HashMap<u32, String>>,
    pub custom_sections: HashMap<String, Vec<u8>>,
}

Container for module data including memories, globals, tables, imports, and exports.

Fields
memories: Map<LocalMemoryIndex, MemoryDescriptor>

Map of memory index to memory descriptors.

globals: Map<LocalGlobalIndex, GlobalInit>

Map of global index to global descriptors.

tables: Map<LocalTableIndex, TableDescriptor>

Map of table index to table descriptors.

imported_functions: Map<ImportedFuncIndex, ImportName>

Map of imported function index to import name.

imported_memories: Map<ImportedMemoryIndex, (ImportName, MemoryDescriptor)>

Map of imported memory index to import name and memory descriptor.

imported_tables: Map<ImportedTableIndex, (ImportName, TableDescriptor)>

Map of imported table index to import name and table descriptor.

imported_globals: Map<ImportedGlobalIndex, (ImportName, GlobalDescriptor)>

Map of imported global index to import name and global descriptor.

exports: IndexMap<String, ExportIndex>

Map of string to export index.

data_initializers: Vec<DataInitializer>

Vector of data initializers.

elem_initializers: Vec<TableInitializer>

Vector of table initializers.

start_func: Option<FuncIndex>

Index of optional start function.

func_assoc: Map<FuncIndex, SigIndex>

Map function index to signature index.

signatures: Map<SigIndex, FuncSig>

Map signature index to function signature.

backend: String

Backend.

namespace_table: StringTable<NamespaceIndex>

Table of namespace indexes.

name_table: StringTable<NameIndex>

Table of name indexes.

em_symbol_map: Option<HashMap<u32, String>>

Symbol information from emscripten.

custom_sections: HashMap<String, Vec<u8>>

Custom sections.

Methods

impl ModuleInfo

pub fn import_custom_sections(&mut self, wasm: &[u8]) -> ParseResult<()>

Creates custom section info from the given wasm file.