Module wasmer_runtime_core::prelude

The prelude module is a helper module used to bring commonly used runtime core imports into scope.

Re-exports

pub use crate::import::ImportObject;

pub use crate::import::Namespace;

pub use crate::types::FuncIndex;

pub use crate::types::GlobalIndex;

pub use crate::types::ImportedFuncIndex;

pub use crate::types::ImportedGlobalIndex;

pub use crate::types::ImportedMemoryIndex;

pub use crate::types::ImportedTableIndex;

pub use crate::types::LocalFuncIndex;

pub use crate::types::LocalGlobalIndex;

pub use crate::types::LocalMemoryIndex;

pub use crate::types::LocalTableIndex;

pub use crate::types::MemoryIndex;

pub use crate::types::TableIndex;

pub use crate::types::Type;

pub use crate::types::Value;

pub use crate::vm;

pub use crate::func;

pub use crate::imports;