Module wasmer_runtime_core::state

The state module is used to track state of a running web assembly instances so that state could read or updated at runtime. Use cases include generating stack traces, switching generated code from one tier to another, or serializing state of a running instace.

Modules

x64_decl
Declarations for x86-64 registers.

x64
The x64 state module contains functions to generate state and code for x64 targets.

Structs

CodeVersion
A CodeVersion is a container for a unit of generated code for a module.

ExecutionStateImage
An image of the execution state.

FunctionStateMap
A map of function states.

InstanceImage
Represents an image of an Instance including its memory, globals, and execution state.

MachineState
A container for the state of a running wasm instance.

MachineStateDiff
A diff of two MachineStates.

ModuleStateMap
A map of module state.

OffsetInfo
Info for an offset.

RegisterIndex
An index to a register

WasmFunctionStateDump
State dump of a wasm function.

Enums

MachineValue
A kind of machine value.

SuspendOffset
A kind of suspend offset.

WasmAbstractValue
A kind of wasm or constant value