Module wasmer_runtime_core::state::x64

The x64 state module contains functions to generate state and code for x64 targets.

Re-exports

pub use super::x64_decl::*;

Functions

build_instance_image
Builds an InstanceImage for the given Ctx and ExecutionStateImage.

invoke_call_return_on_stack⚠
Invokes a call return on the stack for the given module state map, code base, instance image and context.

new_machine_state
Create a new MachineState with default values.

read_stack⚠
Returns a ExecutionStateImage for the given versions, stack, initial registers and initial address.