Struct wasmer_runtime_core::Func

pub struct Func<'a, Args = (), Rets = (), Inner: Kind = Wasm> { /* fields omitted */ }

Methods

impl<'a, Args, Rets> Func<'a, Args, Rets, Host> where
Args: WasmTypeList,
Rets: WasmTypeList,

pub fn new<F>(f: F) -> Func<'a, Args, Rets, Host> where
F: ExternalFunction<Args, Rets>,

impl<'a, Args, Rets, Inner> Func<'a, Args, Rets, Inner> where
Args: WasmTypeList,
Rets: WasmTypeList,
Inner: Kind,

pub fn params(&self) -> &'static [Type]

pub fn returns(&self) -> &'static [Type]

impl<'a, A: WasmExternType, Rets> Func<'a, (A,), Rets, Wasm> where
Rets: WasmTypeList,

pub fn call(&self, a: A) -> Result<Rets, RuntimeError>

impl<'a, Rets> Func<'a, (), Rets, Wasm> where
Rets: WasmTypeList,

pub fn call(&self) -> Result<Rets, RuntimeError>

impl<'a, A: WasmExternType, Rets> Func<'a, A, Rets, Wasm> where
Rets: WasmTypeList,

pub fn call(&self, A: A) -> Result<Rets, RuntimeError>

impl<'a, A: WasmExternType, B: WasmExternType, Rets> Func<'a, (A, B), Rets, Wasm> where
Rets: WasmTypeList,

pub fn call(&self, A: A, B: B) -> Result<Rets, RuntimeError>

impl<'a, A: WasmExternType, B: WasmExternType, C: WasmExternType, Rets> Func<'a, (A, B, C), Rets, Wasm> where
Rets: WasmTypeList,

pub fn call(&self, A: A, B: B, C: C) -> Result<Rets, RuntimeError>

impl<'a, A: WasmExternType, B: WasmExternType, C: WasmExternType, D: WasmExternType, Rets> Func<'a, (A, B, C, D), Rets, Wasm> where
Rets: WasmTypeList,

pub fn call(&self, A: A, B: B, C: C, D: D) -> Result<Rets, RuntimeError>

impl<'a, A: WasmExternType, B: WasmExternType, C: WasmExternType, D: WasmExternType, E: WasmExternType, Rets> Func<'a, (A, B, C, D, E), Rets, Wasm> where
Rets: WasmTypeList,

pub fn call(&self, A: A, B: B, C: C, D: D, E: E) -> Result<Rets, RuntimeError>

impl<'a, A: WasmExternType, B: WasmExternType, C: WasmExternType, D: WasmExternType, E: WasmExternType, F: WasmExternType, Rets> Func<'a, (A, B, C, D, E, F), Rets, Wasm> where
Rets: WasmTypeList,

pub fn call(
&self,
A: A,
B: B,
C: C,
D: D,
E: E,
F: F
) -> Result<Rets, RuntimeError>

impl<'a, A: WasmExternType, B: WasmExternType, C: WasmExternType, D: WasmExternType, E: WasmExternType, F: WasmExternType, G: WasmExternType, Rets> Func<'a, (A, B, C, D, E, F, G), Rets, Wasm> where
Rets: WasmTypeList,

pub fn call(
&self,
A: A,
B: B,
C: C,
D: D,
E: E,
F: F,
G: G
) -> Result<Rets, RuntimeError>

impl<'a, A: WasmExternType, B: WasmExternType, C: WasmExternType, D: WasmExternType, E: WasmExternType, F: WasmExternType, G: WasmExternType, H: WasmExternType, Rets> Func<'a, (A, B, C, D, E, F, G, H), Rets, Wasm> where
Rets: WasmTypeList,

pub fn call(
&self,
A: A,
B: B,
C: C,
D: D,
E: E,
F: F,
G: G,
H: H
) -> Result<Rets, RuntimeError>

impl<'a, A: WasmExternType, B: WasmExternType, C: WasmExternType, D: WasmExternType, E: WasmExternType, F: WasmExternType, G: WasmExternType, H: WasmExternType, I: WasmExternType, Rets> Func<'a, (A, B, C, D, E, F, G, H, I), Rets, Wasm> where
Rets: WasmTypeList,

pub fn call(
&self,
A: A,
B: B,
C: C,
D: D,
E: E,
F: F,
G: G,
H: H,
I: I
) -> Result<Rets, RuntimeError>

impl<'a, A: WasmExternType, B: WasmExternType, C: WasmExternType, D: WasmExternType, E: WasmExternType, F: WasmExternType, G: WasmExternType, H: WasmExternType, I: WasmExternType, J: WasmExternType, Rets> Func<'a, (A, B, C, D, E, F, G, H, I, J), Rets, Wasm> where
Rets: WasmTypeList,

pub fn call(
&self,
A: A,
B: B,
C: C,
D: D,
E: E,
F: F,
G: G,
H: H,
I: I,
J: J
) -> Result<Rets, RuntimeError>

impl<'a, A: WasmExternType, B: WasmExternType, C: WasmExternType, D: WasmExternType, E: WasmExternType, F: WasmExternType, G: WasmExternType, H: WasmExternType, I: WasmExternType, J: WasmExternType, K: WasmExternType, Rets> Func<'a, (A, B, C, D, E, F, G, H, I, J, K), Rets, Wasm> where
Rets: WasmTypeList,

pub fn call(
&self,
A: A,
B: B,
C: C,
D: D,
E: E,
F: F,
G: G,
H: H,
I: I,
J: J,
K: K
) -> Result<Rets, RuntimeError>

impl<'a, A: WasmExternType, B: WasmExternType, C: WasmExternType, D: WasmExternType, E: WasmExternType, F: WasmExternType, G: WasmExternType, H: WasmExternType, I: WasmExternType, J: WasmExternType, K: WasmExternType, L: WasmExternType, Rets> Func<'a, (A, B, C, D, E, F, G, H, I, J, K, L), Rets, Wasm> where
Rets: WasmTypeList,

pub fn call(
&self,
A: A,
B: B,
C: C,
D: D,
E: E,
F: F,
G: G,
H: H,
I: I,
J: J,
K: K,
L: L
) -> Result<Rets, RuntimeError>