Trait wasmer_runtime_core::structures::TypedIndex

pub trait TypedIndex: Copy + Clone {
    fn new(index: usize) -> Self;

    fn index(&self) -> usize;
}

A trait for dealing with type-safe indices into associative data structures like Maps.

Through the use of this trait, we get compile time checks that we are not using the wrong type of index into our data structures.

It acts as a thin wrapper over usize and in most usage patterns has no runtime overhead.

Required methods

fn new(index: usize) -> Self

Create a new instance of TypedIndex from a raw index value

fn index(&self) -> usize

Get the raw index value from the TypedIndex

Loading content...