Module wasmer_runtime_core::table

The runtime table module contains data structures and functions used to create and update wasm tables.

Structs

Anyfunc
Anyfunc data type.

Table
Container with a descriptor and a reference to a table storage.

Enums

Element
Kind of table element.

TableStorage
Kind of table storage.