Module wasmer_runtime_core::trampoline_x64

Trampoline generator for carrying context with function pointer.

This makes use of the mm0 register to pass the context as an implicit "parameter" because mm0 is not used to pass parameters and is almost never used by modern compilers. It's still better to call get_context() as early as possible in the callee function though, as a good practice.

Variadic functions are not supported because rax is used by the trampoline code.

Structs

TrampolineBuffer
Trampoline Buffer.

TrampolineBufferBuilder
Trampoline Buffer Builder.

Enums

CallContext
An opaque type for context pointers.

CallTarget
An opaque type for pointers to a callable memory location.

Trampoline
An opaque type for generated trampolines' call entries.

Functions

get_context
Calls GET_CONTEXT and returns the current context.