Struct wasmer_runtime_core::typed_func::DynamicFunc

pub struct DynamicFunc<'a> {
    _inner: Box<dyn Kind>,
    func: NonNull<Func>,
    func_env: Option<NonNull<FuncEnv>>,
    vmctx: *mut Ctx,
    signature: Arc<FuncSig>,
    _phantom: PhantomData<&'a ()>,
}

Represents a type-erased function provided by either the host or the WebAssembly program.

Fields
_inner: Box<dyn Kind>func: NonNull<Func>

The function pointer.

func_env: Option<NonNull<FuncEnv>>

The function environment.

vmctx: *mut Ctx

The famous vm::Ctx.

signature: Arc<FuncSig>

The runtime signature of this function.

When converted from a Func, this is determined by the static Args and Rets type parameters. otherwise the signature is dynamically assigned during DynamicFunc creation, usually when creating a polymorphic host function.

_phantom: PhantomData<&'a ()>
Methods

impl<'a> DynamicFunc<'a>

pub fn new<F>(signature: Arc<FuncSig>, func: F) -> Self where
F: Fn(&mut Ctx, &[Value]) -> Vec<Value> + 'static,

Creates a dynamic function that is polymorphic over its argument and return types.