Trait wasmer_runtime_core::typed_func::ExternalFunctionKind

pub trait ExternalFunctionKind { }

Empty trait to specify the kind of ExternalFunction: With or without a vm::Ctx argument. See the ExplicitVmCtx and the ImplicitVmCtx structures.

This type is never aimed to be used by a user. It is used by the trait system to automatically generate an appropriate wrap function.