Trait wasmer_runtime_core::typed_func::Kind

pub trait Kind { }

This is just an empty trait to constrict that types that can be put into the third/fourth (depending if you include lifetimes) of the Func struct.