Trait wasmer_runtime_core::typed_func::WasmTypeList

pub trait WasmTypeList {
    type CStruct;
    type RetArray: AsMut<[u64]>;
    fn from_ret_array(array: Self::RetArray) -> Self;

    fn empty_ret_array() -> Self::RetArray;

    fn from_c_struct(c_struct: Self::CStruct) -> Self;

    fn into_c_struct(self) -> Self::CStruct;

    fn types() -> &'static [Type];

    unsafe fn call<Rets>(
        self, 
        f: NonNull<Func>, 
        wasm: Wasm, 
        ctx: *mut Ctx
    ) -> Result<Rets, RuntimeError>
    where
        Rets: WasmTypeList;
}

Represents a list of WebAssembly values.

Associated Types

type CStruct

CStruct type.

type RetArray: AsMut<[u64]>

Array of return values.

Loading content...
Required methods

fn from_ret_array(array: Self::RetArray) -> Self

Construct Self based on an array of returned values.

fn empty_ret_array() -> Self::RetArray

Generates an empty array that will hold the returned values of the WebAssembly function.

fn from_c_struct(c_struct: Self::CStruct) -> Self

Transforms C values into Rust values.

fn into_c_struct(self) -> Self::CStruct

Transforms Rust values into C values.

fn types() -> &'static [Type]

Get types of the current values.

unsafe fn call<Rets>(
self,
f: NonNull<Func>,
wasm: Wasm,
ctx: *mut Ctx
) -> Result<Rets, RuntimeError> where
Rets: WasmTypeList,

This method is used to distribute the values onto a function, e.g. (1, 2).call(func, …). This form is unlikely to be used directly in the code, see the Func:call implementation.

Loading content...