Module wasmer_runtime_core::units

The units module provides common WebAssembly units like Pages and conversion functions into other units.

Structs

Bytes
Units of WebAssembly memory in terms of 8-bit bytes.

Pages
Units of WebAssembly pages (as specified to be 65,536 bytes).

Constants

WASM_MAX_PAGES
Tbe max number of wasm pages allowed.

WASM_MIN_PAGES
The minimum number of wasm pages allowed.

WASM_PAGE_SIZE
The page size in bytes of a wasm page.