Module wasmer_runtime_core::vm

The runtime vm module contains data structures and helper functions used during runtime to execute wasm instance functions.

Structs

Anyfunc
Caller-checked anyfunc

Ctx
The context of the currently running WebAssembly instance.

Func
Represents a function pointer. It is mostly used in the typed_func module within the wrap functions, to wrap imported functions.

FuncCtx
Represents a function context. It is used by imported functions only.

FuncEnv
Represents a function environment pointer, like a captured environment of a closure. It is mostly used in the typed_func module within the wrap functions, to wrap imported functions.

ImportBacking
The ImportBacking stores references to the imported resources of an Instance. This includes imported memories, tables, globals and functions.

ImportedFunc
An imported function is a function pointer associated to a function context.

InternalField
An internal field.

Intrinsics
A container for VM instrinsic functions

LocalBacking
The LocalBacking "owns" the memory used by all the local resources of an Instance. That is, local memories, tables, and globals (as well as some additional data for the virtual call machinery).

LocalGlobal
Definition of a global used by the VM.

LocalMemory
Definition of a memory used by the VM.

LocalTable
Definition of a table used by the VM. (obviously)

SigId
Identifier for a function signature.

Constants

INTERNALS_SIZE
Size of the array for internal instance usage

Statics

INTRINSICS_IMPORTED_DYNAMIC_MEMORY
Imported dynamic memory intrinsics

INTRINSICS_IMPORTED_STATIC_MEMORY
Imported static memory intrinsics

INTRINSICS_LOCAL_DYNAMIC_MEMORY
Local dynamic memory intrinsics

INTRINSICS_LOCAL_STATIC_MEMORY
Local static memory intrinsics