Struct wasmer_runtime_core::vm::InternalCtx

#[repr(C)]
pub struct InternalCtx {
    pub memories: *mut *mut LocalMemory,
    pub tables: *mut *mut LocalTable,
    pub globals: *mut *mut LocalGlobal,
    pub imported_memories: *mut *mut LocalMemory,
    pub imported_tables: *mut *mut LocalTable,
    pub imported_globals: *mut *mut LocalGlobal,
    pub imported_funcs: *mut ImportedFunc,
    pub dynamic_sigindices: *const SigId,
    pub intrinsics: *const Intrinsics,
    pub stack_lower_bound: *mut u8,
    pub memory_base: *mut u8,
    pub memory_bound: usize,
    pub internals: *mut [u64; 256],
    pub interrupt_signal_mem: *mut u8,
}

The internal context of the currently running WebAssembly instance.

Fields
memories: *mut *mut LocalMemory

A pointer to an array of locally-defined memories, indexed by MemoryIndex.

tables: *mut *mut LocalTable

A pointer to an array of locally-defined tables, indexed by TableIndex.

globals: *mut *mut LocalGlobal

A pointer to an array of locally-defined globals, indexed by GlobalIndex.

imported_memories: *mut *mut LocalMemory

A pointer to an array of imported memories, indexed by MemoryIndex,

imported_tables: *mut *mut LocalTable

A pointer to an array of imported tables, indexed by TableIndex.

imported_globals: *mut *mut LocalGlobal

A pointer to an array of imported globals, indexed by GlobalIndex.

imported_funcs: *mut ImportedFunc

A pointer to an array of imported functions, indexed by FuncIndex.

dynamic_sigindices: *const SigId

A pointer to an array of signature ids. Conceptually, this maps from a static, module-local signature id to a runtime-global signature id. This is used to allow call-indirect to other modules safely.

intrinsics: *const Intrinsics

Const pointer to Intrinsics.

stack_lower_bound: *mut u8

Stack lower bound.

memory_base: *mut u8

Mutable pointer to memory base.

memory_bound: usize

Memory bound.

internals: *mut [u64; 256]

Mutable pointer to internal fields.

interrupt_signal_mem: *mut u8

Interrupt signal mem.