Struct wasmer_runtime_core::Func

pub struct Func<'a, Args = (), Rets = (), Inner: Kind = Wasm> { /* fields omitted */ }

Represents a function that can be used by WebAssembly.

Methods

impl<'a, Args, Rets> Func<'a, Args, Rets, Host> where
Args: WasmTypeList,
Rets: WasmTypeList,

pub fn new<F, Kind>(func: F) -> Func<'a, Args, Rets, Host> where
Kind: ExternalFunctionKind,
F: ExternalFunction<Kind, Args, Rets>,

Creates a new Func.

impl<'a, Args, Rets, Inner> Func<'a, Args, Rets, Inner> where
Args: WasmTypeList,
Rets: WasmTypeList,
Inner: Kind,

pub fn params(&self) -> &'static [Type]

Returns the types of the function inputs.

pub fn returns(&self) -> &'static [Type]

Returns the types of the function outputs.

pub fn get_vm_func(&self) -> NonNull<Func>

Get the underlying func pointer.

impl<'a, Rets> Func<'a, (), Rets, Wasm> where
Rets: WasmTypeList,

pub fn call(&self) -> Result<Rets, RuntimeError>

Call the typed func and return results.

impl<'a, A, Rets> Func<'a, A, Rets, Wasm> where
A: WasmExternType,
Rets: WasmTypeList,

pub fn call(&self, A: A) -> Result<Rets, RuntimeError>

Call the typed func and return results.

impl<'a, A, B, Rets> Func<'a, (A, B), Rets, Wasm> where
A: WasmExternType,
B: WasmExternType,
Rets: WasmTypeList,

pub fn call(&self, A: A, B: B) -> Result<Rets, RuntimeError>

Call the typed func and return results.

impl<'a, A, B, C, Rets> Func<'a, (A, B, C), Rets, Wasm> where
A: WasmExternType,
B: WasmExternType,
C: WasmExternType,
Rets: WasmTypeList,

pub fn call(&self, A: A, B: B, C: C) -> Result<Rets, RuntimeError>

Call the typed func and return results.

impl<'a, A, B, C, D, Rets> Func<'a, (A, B, C, D), Rets, Wasm> where
A: WasmExternType,
B: WasmExternType,
C: WasmExternType,
D: WasmExternType,
Rets: WasmTypeList,

pub fn call(&self, A: A, B: B, C: C, D: D) -> Result<Rets, RuntimeError>

Call the typed func and return results.

impl<'a, A, B, C, D, E, Rets> Func<'a, (A, B, C, D, E), Rets, Wasm> where
A: WasmExternType,
B: WasmExternType,
C: WasmExternType,
D: WasmExternType,
E: WasmExternType,
Rets: WasmTypeList,

pub fn call(&self, A: A, B: B, C: C, D: D, E: E) -> Result<Rets, RuntimeError>

Call the typed func and return results.

impl<'a, A, B, C, D, E, F, Rets> Func<'a, (A, B, C, D, E, F), Rets, Wasm> where
A: WasmExternType,
B: WasmExternType,
C: WasmExternType,
D: WasmExternType,
E: WasmExternType,
F: WasmExternType,
Rets: WasmTypeList,

pub fn call(
&self,
A: A,
B: B,
C: C,
D: D,
E: E,
F: F
) -> Result<Rets, RuntimeError>

Call the typed func and return results.

impl<'a, A, B, C, D, E, F, G, Rets> Func<'a, (A, B, C, D, E, F, G), Rets, Wasm> where
A: WasmExternType,
B: WasmExternType,
C: WasmExternType,
D: WasmExternType,
E: WasmExternType,
F: WasmExternType,
G: WasmExternType,
Rets: WasmTypeList,

pub fn call(
&self,
A: A,
B: B,
C: C,
D: D,
E: E,
F: F,
G: G
) -> Result<Rets, RuntimeError>

Call the typed func and return results.

impl<'a, A, B, C, D, E, F, G, H, Rets> Func<'a, (A, B, C, D, E, F, G, H), Rets, Wasm> where
A: WasmExternType,
B: WasmExternType,
C: WasmExternType,
D: WasmExternType,
E: WasmExternType,
F: WasmExternType,
G: WasmExternType,
H: WasmExternType,
Rets: WasmTypeList,

pub fn call(
&self,
A: A,
B: B,
C: C,
D: D,
E: E,
F: F,
G: G,
H: H
) -> Result<Rets, RuntimeError>

Call the typed func and return results.

impl<'a, A, B, C, D, E, F, G, H, I, Rets> Func<'a, (A, B, C, D, E, F, G, H, I), Rets, Wasm> where
A: WasmExternType,
B: WasmExternType,
C: WasmExternType,
D: WasmExternType,
E: WasmExternType,
F: WasmExternType,
G: WasmExternType,
H: WasmExternType,
I: WasmExternType,
Rets: WasmTypeList,

pub fn call(
&self,
A: A,
B: B,
C: C,
D: D,
E: E,
F: F,
G: G,
H: H,
I: I
) -> Result<Rets, RuntimeError>

Call the typed func and return results.

impl<'a, A, B, C, D, E, F, G, H, I, J, Rets> Func<'a, (A, B, C, D, E, F, G, H, I, J), Rets, Wasm> where
A: WasmExternType,
B: WasmExternType,
C: WasmExternType,
D: WasmExternType,
E: WasmExternType,
F: WasmExternType,
G: WasmExternType,
H: WasmExternType,
I: WasmExternType,
J: WasmExternType,
Rets: WasmTypeList,

pub fn call(
&self,
A: A,
B: B,
C: C,
D: D,
E: E,
F: F,
G: G,
H: H,
I: I,
J: J
) -> Result<Rets, RuntimeError>

Call the typed func and return results.

impl<'a, A, B, C, D, E, F, G, H, I, J, K, Rets> Func<'a, (A, B, C, D, E, F, G, H, I, J, K), Rets, Wasm> where
A: WasmExternType,
B: WasmExternType,
C: WasmExternType,
D: WasmExternType,
E: WasmExternType,
F: WasmExternType,
G: WasmExternType,
H: WasmExternType,
I: WasmExternType,
J: WasmExternType,
K: WasmExternType,
Rets: WasmTypeList,

pub fn call(
&self,
A: A,
B: B,
C: C,
D: D,
E: E,
F: F,
G: G,
H: H,
I: I,
J: J,
K: K
) -> Result<Rets, RuntimeError>

Call the typed func and return results.

impl<'a, A, B, C, D, E, F, G, H, I, J, K, L, Rets> Func<'a, (A, B, C, D, E, F, G, H, I, J, K, L), Rets, Wasm> where
A: WasmExternType,
B: WasmExternType,
C: WasmExternType,
D: WasmExternType,
E: WasmExternType,
F: WasmExternType,
G: WasmExternType,
H: WasmExternType,
I: WasmExternType,
J: WasmExternType,
K: WasmExternType,
L: WasmExternType,
Rets: WasmTypeList,

pub fn call(
&self,
A: A,
B: B,
C: C,
D: D,
E: E,
F: F,
G: G,
H: H,
I: I,
J: J,
K: K,
L: L
) -> Result<Rets, RuntimeError>

Call the typed func and return results.

impl<'a, A, B, C, D, E, F, G, H, I, J, K, L, M, Rets> Func<'a, (A, B, C, D, E, F, G, H, I, J, K, L, M), Rets, Wasm> where
A: WasmExternType,
B: WasmExternType,
C: WasmExternType,
D: WasmExternType,
E: WasmExternType,
F: WasmExternType,
G: WasmExternType,
H: WasmExternType,
I: WasmExternType,
J: WasmExternType,
K: WasmExternType,
L: WasmExternType,
M: WasmExternType,
Rets: WasmTypeList,

pub fn call(
&self,
A: A,
B: B,
C: C,
D: D,
E: E,
F: F,
G: G,
H: H,
I: I,
J: J,
K: K,
L: L,
M: M
) -> Result<Rets, RuntimeError>

Call the typed func and return results.

impl<'a, A, B, C, D, E, F, G, H, I, J, K, L, M, N, Rets> Func<'a, (A, B, C, D, E, F, G, H, I, J, K, L, M, N), Rets, Wasm> where
A: WasmExternType,
B: WasmExternType,
C: WasmExternType,
D: WasmExternType,
E: WasmExternType,
F: WasmExternType,
G: WasmExternType,
H: WasmExternType,
I: WasmExternType,
J: WasmExternType,
K: WasmExternType,
L: WasmExternType,
M: WasmExternType,
N: WasmExternType,
Rets: WasmTypeList,

pub fn call(
&self,
A: A,
B: B,
C: C,
D: D,
E: E,
F: F,
G: G,
H: H,
I: I,
J: J,
K: K,
L: L,
M: M,
N: N
) -> Result<Rets, RuntimeError>

Call the typed func and return results.

impl<'a, A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, Rets> Func<'a, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O), Rets, Wasm> where
A: WasmExternType,
B: WasmExternType,
C: WasmExternType,
D: WasmExternType,
E: WasmExternType,
F: WasmExternType,
G: WasmExternType,
H: WasmExternType,
I: WasmExternType,
J: WasmExternType,
K: WasmExternType,
L: WasmExternType,
M: WasmExternType,
N: WasmExternType,
O: WasmExternType,
Rets: WasmTypeList,

pub fn call(
&self,
A: A,
B: B,
C: C,
D: D,
E: E,
F: F,
G: G,
H: H,
I: I,
J: J,
K: K,
L: L,
M: M,
N: N,
O: O
) -> Result<Rets, RuntimeError>

Call the typed func and return results.

impl<'a, A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Rets> Func<'a, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P), Rets, Wasm> where
A: WasmExternType,
B: WasmExternType,
C: WasmExternType,
D: WasmExternType,
E: WasmExternType,
F: WasmExternType,
G: WasmExternType,
H: WasmExternType,
I: WasmExternType,
J: WasmExternType,
K: WasmExternType,
L: WasmExternType,
M: WasmExternType,
N: WasmExternType,
O: WasmExternType,
P: WasmExternType,
Rets: WasmTypeList,

pub fn call(
&self,
A: A,
B: B,
C: C,
D: D,
E: E,
F: F,
G: G,
H: H,
I: I,
J: J,
K: K,
L: L,
M: M,
N: N,
O: O,
P: P
) -> Result<Rets, RuntimeError>

Call the typed func and return results.

impl<'a, A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, Rets> Func<'a, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q), Rets, Wasm> where
A: WasmExternType,
B: WasmExternType,
C: WasmExternType,
D: WasmExternType,
E: WasmExternType,
F: WasmExternType,
G: WasmExternType,
H: WasmExternType,
I: WasmExternType,
J: WasmExternType,
K: WasmExternType,
L: WasmExternType,
M: WasmExternType,
N: WasmExternType,
O: WasmExternType,
P: WasmExternType,
Q: WasmExternType,
Rets: WasmTypeList,

pub fn call(
&self,
A: A,
B: B,
C: C,
D: D,
E: E,
F: F,
G: G,
H: H,
I: I,
J: J,
K: K,
L: L,
M: M,
N: N,
O: O,
P: P,
Q: Q
) -> Result<Rets, RuntimeError>

Call the typed func and return results.

impl<'a, A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, R, Rets> Func<'a, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, R), Rets, Wasm> where
A: WasmExternType,
B: WasmExternType,
C: WasmExternType,
D: WasmExternType,
E: WasmExternType,
F: WasmExternType,
G: WasmExternType,
H: WasmExternType,
I: WasmExternType,
J: WasmExternType,
K: WasmExternType,
L: WasmExternType,
M: WasmExternType,
N: WasmExternType,
O: WasmExternType,
P: WasmExternType,
Q: WasmExternType,
R: WasmExternType,
Rets: WasmTypeList,

pub fn call(
&self,
A: A,
B: B,
C: C,
D: D,
E: E,
F: F,
G: G,
H: H,
I: I,
J: J,
K: K,
L: L,
M: M,
N: N,
O: O,
P: P,
Q: Q,
R: R
) -> Result<Rets, RuntimeError>

Call the typed func and return results.

impl<'a, A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, R, S, Rets> Func<'a, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, R, S), Rets, Wasm> where
A: WasmExternType,
B: WasmExternType,
C: WasmExternType,
D: WasmExternType,
E: WasmExternType,
F: WasmExternType,
G: WasmExternType,
H: WasmExternType,
I: WasmExternType,
J: WasmExternType,
K: WasmExternType,
L: WasmExternType,
M: WasmExternType,
N: WasmExternType,
O: WasmExternType,
P: WasmExternType,
Q: WasmExternType,
R: WasmExternType,
S: WasmExternType,
Rets: WasmTypeList,

pub fn call(
&self,
A: A,
B: B,
C: C,
D: D,
E: E,
F: F,
G: G,
H: H,
I: I,
J: J,
K: K,
L: L,
M: M,
N: N,
O: O,
P: P,
Q: Q,
R: R,
S: S
) -> Result<Rets, RuntimeError>

Call the typed func and return results.

impl<'a, A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, R, S, T, Rets> Func<'a, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, R, S, T), Rets, Wasm> where
A: WasmExternType,
B: WasmExternType,
C: WasmExternType,
D: WasmExternType,
E: WasmExternType,
F: WasmExternType,
G: WasmExternType,
H: WasmExternType,
I: WasmExternType,
J: WasmExternType,
K: WasmExternType,
L: WasmExternType,
M: WasmExternType,
N: WasmExternType,
O: WasmExternType,
P: WasmExternType,
Q: WasmExternType,
R: WasmExternType,
S: WasmExternType,
T: WasmExternType,
Rets: WasmTypeList,

pub fn call(
&self,
A: A,
B: B,
C: C,
D: D,
E: E,
F: F,
G: G,
H: H,
I: I,
J: J,
K: K,
L: L,
M: M,
N: N,
O: O,
P: P,
Q: Q,
R: R,
S: S,
T: T
) -> Result<Rets, RuntimeError>

Call the typed func and return results.

impl<'a, A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, R, S, T, U, Rets> Func<'a, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, R, S, T, U), Rets, Wasm> where
A: WasmExternType,
B: WasmExternType,
C: WasmExternType,
D: WasmExternType,
E: WasmExternType,
F: WasmExternType,
G: WasmExternType,
H: WasmExternType,
I: WasmExternType,
J: WasmExternType,
K: WasmExternType,
L: WasmExternType,
M: WasmExternType,
N: WasmExternType,
O: WasmExternType,
P: WasmExternType,
Q: WasmExternType,
R: WasmExternType,
S: WasmExternType,
T: WasmExternType,
U: WasmExternType,
Rets: WasmTypeList,

pub fn call(
&self,
A: A,
B: B,
C: C,
D: D,
E: E,
F: F,
G: G,
H: H,
I: I,
J: J,
K: K,
L: L,
M: M,
N: N,
O: O,
P: P,
Q: Q,
R: R,
S: S,
T: T,
U: U
) -> Result<Rets, RuntimeError>

Call the typed func and return results.

impl<'a, A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, R, S, T, U, V, Rets> Func<'a, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, R, S, T, U, V), Rets, Wasm> where
A: WasmExternType,
B: WasmExternType,
C: WasmExternType,
D: WasmExternType,
E: WasmExternType,
F: WasmExternType,
G: WasmExternType,
H: WasmExternType,
I: WasmExternType,
J: WasmExternType,
K: WasmExternType,
L: WasmExternType,
M: WasmExternType,
N: WasmExternType,
O: WasmExternType,
P: WasmExternType,
Q: WasmExternType,
R: WasmExternType,
S: WasmExternType,
T: WasmExternType,
U: WasmExternType,
V: WasmExternType,
Rets: WasmTypeList,

pub fn call(
&self,
A: A,
B: B,
C: C,
D: D,
E: E,
F: F,
G: G,
H: H,
I: I,
J: J,
K: K,
L: L,
M: M,
N: N,
O: O,
P: P,
Q: Q,
R: R,
S: S,
T: T,
U: U,
V: V
) -> Result<Rets, RuntimeError>

Call the typed func and return results.

impl<'a, A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, R, S, T, U, V, W, Rets> Func<'a, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, R, S, T, U, V, W), Rets, Wasm> where
A: WasmExternType,
B: WasmExternType,
C: WasmExternType,
D: WasmExternType,
E: WasmExternType,
F: WasmExternType,
G: WasmExternType,
H: WasmExternType,
I: WasmExternType,
J: WasmExternType,
K: WasmExternType,
L: WasmExternType,
M: WasmExternType,
N: WasmExternType,
O: WasmExternType,
P: WasmExternType,
Q: WasmExternType,
R: WasmExternType,
S: WasmExternType,
T: WasmExternType,
U: WasmExternType,
V: WasmExternType,
W: WasmExternType,
Rets: WasmTypeList,

pub fn call(
&self,
A: A,
B: B,
C: C,
D: D,
E: E,
F: F,
G: G,
H: H,
I: I,
J: J,
K: K,
L: L,
M: M,
N: N,
O: O,
P: P,
Q: Q,
R: R,
S: S,
T: T,
U: U,
V: V,
W: W
) -> Result<Rets, RuntimeError>

Call the typed func and return results.

impl<'a, A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, R, S, T, U, V, W, X, Rets> Func<'a, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, R, S, T, U, V, W, X), Rets, Wasm> where
A: WasmExternType,
B: WasmExternType,
C: WasmExternType,
D: WasmExternType,
E: WasmExternType,
F: WasmExternType,
G: WasmExternType,
H: WasmExternType,
I: WasmExternType,
J: WasmExternType,
K: WasmExternType,
L: WasmExternType,
M: WasmExternType,
N: WasmExternType,
O: WasmExternType,
P: WasmExternType,
Q: WasmExternType,
R: WasmExternType,
S: WasmExternType,
T: WasmExternType,
U: WasmExternType,
V: WasmExternType,
W: WasmExternType,
X: WasmExternType,
Rets: WasmTypeList,

pub fn call(
&self,
A: A,
B: B,
C: C,
D: D,
E: E,
F: F,
G: G,
H: H,
I: I,
J: J,
K: K,
L: L,
M: M,
N: N,
O: O,
P: P,
Q: Q,
R: R,
S: S,
T: T,
U: U,
V: V,
W: W,
X: X
) -> Result<Rets, RuntimeError>

Call the typed func and return results.

impl<'a, A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, R, S, T, U, V, W, X, Y, Rets> Func<'a, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, R, S, T, U, V, W, X, Y), Rets, Wasm> where
A: WasmExternType,
B: WasmExternType,
C: WasmExternType,
D: WasmExternType,
E: WasmExternType,
F: WasmExternType,
G: WasmExternType,
H: WasmExternType,
I: WasmExternType,
J: WasmExternType,
K: WasmExternType,
L: WasmExternType,
M: WasmExternType,
N: WasmExternType,
O: WasmExternType,
P: WasmExternType,
Q: WasmExternType,
R: WasmExternType,
S: WasmExternType,
T: WasmExternType,
U: WasmExternType,
V: WasmExternType,
W: WasmExternType,
X: WasmExternType,
Y: WasmExternType,
Rets: WasmTypeList,

pub fn call(
&self,
A: A,
B: B,
C: C,
D: D,
E: E,
F: F,
G: G,
H: H,
I: I,
J: J,
K: K,
L: L,
M: M,
N: N,
O: O,
P: P,
Q: Q,
R: R,
S: S,
T: T,
U: U,
V: V,
W: W,
X: X,
Y: Y
) -> Result<Rets, RuntimeError>

Call the typed func and return results.

impl<'a, A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, R, S, T, U, V, W, X, Y, Z, Rets> Func<'a, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, R, S, T, U, V, W, X, Y, Z), Rets, Wasm> where
A: WasmExternType,
B: WasmExternType,
C: WasmExternType,
D: WasmExternType,
E: WasmExternType,
F: WasmExternType,
G: WasmExternType,
H: WasmExternType,
I: WasmExternType,
J: WasmExternType,
K: WasmExternType,
L: WasmExternType,
M: WasmExternType,
N: WasmExternType,
O: WasmExternType,
P: WasmExternType,
Q: WasmExternType,
R: WasmExternType,
S: WasmExternType,
T: WasmExternType,
U: WasmExternType,
V: WasmExternType,
W: WasmExternType,
X: WasmExternType,
Y: WasmExternType,
Z: WasmExternType,
Rets: WasmTypeList,

pub fn call(
&self,
A: A,
B: B,
C: C,
D: D,
E: E,
F: F,
G: G,
H: H,
I: I,
J: J,
K: K,
L: L,
M: M,
N: N,
O: O,
P: P,
Q: Q,
R: R,
S: S,
T: T,
U: U,
V: V,
W: W,
X: X,
Y: Y,
Z: Z
) -> Result<Rets, RuntimeError>

Call the typed func and return results.