Trait wasmer_runtime_core::structures::TypedIndex

pub trait TypedIndex: Copy + Clone { }

A trait for dealing with type-safe indices into associative data structures like Maps.

Through the use of this trait, we get compile time checks that we are not using the wrong type of index into our data structures.

It acts as a thin wrapper over usize and in most usage patterns has no runtime overhead.